Enhancement: Codility training lesson EquiLeader
Leader of the vector is found by using stack. Notice that if two
different numbers are removed from vector, the leader does not
change. Use stack to compare last and current number, and
discard both of them if they are different.One by one non dominant
numbers will be eliminated and stack will be full of leader num.

Resolves: #25