Enhancement: Hackerrank challenge AbsoluteElementSums
Hackerrank challenge requires performing addition of giiven number
to elements of an input vector, and then displaying thesum of
absolute value of it's elements. First input vector is sorted,
and zeroboundary between negative and positive numbers are found.
Subsequent queries are performed on initial sum, and this zero
crossing boundary point.

Resolves: #65