Enhancement: Hackerrank challenge FindingTheRunningMedian
Median is calculated with keeping two multisets with one
containing higher numbers, and one containing smaller numbers.
Each new incommer is inserted to the smaller dimensioned set, and
then highest value of lower set is compared with lowest value of
higher set in order to check a swap is needed or not. Median is
easly calculated using highest value of lower set and lowest value
of higher set.

Resolves: #96